Update base.py
The update allows several Arduino IDE versions to be selectable from  the Arduino->Preferences->Select Arduino Application Folder in Mac OS. For example, Arduino IDE 1.0.5 and Arduino IDE 1.5.8 can be selected. In some cases, the two arduino IDE versions are required due to custom board inclusions that are not compatible with both versions.